Refactor and move tests together
